+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Fixed
+- check-log.rb: pattern is now required via option config
+- check-log.rb: update descriptions of crit and warn options
 
 ## [1.3.1] - 2017-10-09
 ### Fixed